refactor(analysis): rearrange stuff in/out of advance settings
