+++ lib/components/analysis/__tests__/profile-request-editor.js
@@ -1,4 +1,5 @@
 // @flow
+
 import React from 'react'
 
 import ProfileRequestEditor from '../profile-request-editor'
@@ -9,35 +10,23 @@ describe('Components > Analysis > Profile Request Editor', () => {
     // nb using enzyme b/c react shallow rendering doesn't play nice with the date picker
     const {snapshot} = mockWithProvider(
       <ProfileRequestEditor
+        disabled={false}
         profileRequest={mockProfileRequest}
-        updateProfileRequest={jest.fn()}
+        setProfileRequest={jest.fn()}
       />
     )
-
-    expect(snapshot()).toMatchSnapshot()
-  })
-
-  it('should render correctly with expanded advanced settings', () => {
-    // nb using enzyme b/c react shallow rendering doesn't play nice with the date picker
-    const {snapshot, wrapper} = mockWithProvider(
-      <ProfileRequestEditor
-        profileRequest={mockProfileRequest}
-        updateProfileRequest={jest.fn()}
-      />
-    )
-
-    wrapper.find('section a').simulate('click')
 
     expect(snapshot()).toMatchSnapshot()
   })
 
   it('should react to input correctly', () => {
-    const updateProfileRequest = jest.fn()
+    const setProfileRequest = jest.fn()
 
     const {wrapper} = mockWithProvider(
       <ProfileRequestEditor
+        disabled={false}
         profileRequest={mockProfileRequest}
-        updateProfileRequest={updateProfileRequest}
+        setProfileRequest={setProfileRequest}
       />
     )
 
@@ -47,15 +36,9 @@ describe('Components > Analysis > Profile Request Editor', () => {
     wrapper.find('input[name="fromTime"]').simulate('change', ch('01:00'))
     wrapper.find('input[name="toTime"]').simulate('change', ch('13:00'))
     // TODO date picker
-    // open advanced settings
-    wrapper.find('section a').simulate('click')
-    wrapper.find('input[name="maxTransfers"]').simulate('change', ch('5'))
-    wrapper.find('input[name="monteCarloDraws"]').simulate('change', ch('319'))
 
-    expect(updateProfileRequest.mock.calls[0][0].fromTime).toEqual(1 * 3600) // 1 AM is 3600s after midnight
-    expect(updateProfileRequest.mock.calls[1][0].toTime).toEqual(13 * 3600)
-    expect(updateProfileRequest.mock.calls[2][0].maxRides).toEqual(6) // five transfers is six rides
-    expect(updateProfileRequest.mock.calls[3][0].monteCarloDraws).toEqual(319)
+    expect(setProfileRequest.mock.calls[0][0].fromTime).toEqual(1 * 3600) // 1 AM is 3600s after midnight
+    expect(setProfileRequest.mock.calls[1][0].toTime).toEqual(13 * 3600)
   })
 })
