+++ lib/components/analysis/index.js
@@ -1,23 +1,21 @@
 // @flow
 import Icon from '@conveyal/woonerf/components/icon'
 import Pure from '@conveyal/woonerf/components/pure'
-import {Map as LeafletMap} from 'leaflet'
 import React from 'react'
 import Select from 'react-select'
-import {sprintf} from 'sprintf-js'
 
-import BookmarkChooser from '../../containers/bookmark-chooser'
-import messages from '../../utils/messages'
+import AdvancedSettings from './advanced-settings'
 import {Button, Group as ButtonGroup} from '../buttons'
-import InnerDock from '../inner-dock'
+import Collapsible from '../collapsible'
+import BookmarkChooser from '../../containers/bookmark-chooser'
 import StackedPercentileSelector
   from '../../containers/stacked-percentile-selector'
+import InnerDock from '../inner-dock'
 import {Group, Slider} from '../input'
-import Collapsible from '../collapsible'
+import OpportunityDatasets from '../../modules/opportunity-datasets'
 import ProfileRequestEditor from './profile-request-editor'
 import ProjectApplicationError from './project-application-error'
-
-import OpportunityDatasets from '../../modules/opportunity-datasets'
+import messages from '../../utils/messages'
 
 import type {
   Modification,
@@ -89,7 +87,6 @@ type Props = {
 
 type State = {
   creatingRegionalAnalysis: boolean,
-  editingBounds: boolean,
   regionalAnalysisName: void | string
 }
 
@@ -102,7 +99,6 @@ export default class SinglePointAnalysis extends Pure {
 
   state = {
     creatingRegionalAnalysis: false,
-    editingBounds: false,
     regionalAnalysisName: undefined
   }
 
@@ -176,37 +172,12 @@ export default class SinglePointAnalysis extends Pure {
     this.setState({creatingRegionalAnalysis: true})
   }
 
-  _showBoundsSelector = () => {
-    this.props.addEditRegionalAnalysisBoundsLayerToMap()
-    this.setState({editingBounds: true})
-  }
-
-  _hideBoundsSelector = () => {
-    this.props.removeEditRegionalAnalysisBoundsLayerFromMap()
-    this.setState({editingBounds: false})
-  }
-
   _setTravelTimePercentile = (e: Event & {target: HTMLInputElement}) =>
     this.props.setProfileRequest({travelTimePercentile: Number(e.target.value)})
 
-  _setRegionalAnalysisBounds = (e: {value: string}) => {
-    const {
-      regionBounds,
-      regionalAnalyses
-    } = this.props
-
-    if (e.value === '__REGION') {
-      this.props.setProfileRequest({bounds: regionBounds})
-    } else if (regionalAnalyses) {
-      const foundAnalyses = regionalAnalyses.find(r => r._id === e.value)
-      if (foundAnalyses) {
-        this.props.setProfileRequest({bounds: webMercatorBoundsToGeographic(foundAnalyses)})
-      }
-    }
-  }
-
   render () {
     const {
+      addEditRegionalAnalysisBoundsLayerToMap,
       downloadComparisonIsochrone,
       downloadIsochrone,
       fetchTravelTimeSurface,
@@ -217,6 +188,10 @@ export default class SinglePointAnalysis extends Pure {
       profileRequest,
       projectApplicationErrors,
       projectApplicationWarnings,
+      regionalAnalyses,
+      regionBounds,
+      removeEditRegionalAnalysisBoundsLayerFromMap,
+      setProfileRequest,
       variants
     } = this.props
     const {creatingRegionalAnalysis} = this.state
@@ -338,10 +313,8 @@ export default class SinglePointAnalysis extends Pure {
             <ProfileRequestEditor
               disabled={disableInputs}
               profileRequest={profileRequest}
-              updateProfileRequest={this.props.setProfileRequest}
+              updateProfileRequest={setProfileRequest}
             />
-
-            {this.renderRegionalAnalysisBoundsSelector(disableInputs)}
 
             {/*
               * A slider for selecting the percentile to use in a regional analysis.
@@ -359,83 +332,23 @@ export default class SinglePointAnalysis extends Pure {
               label={`${messages.analysis.travelTimePercentile} (Regional only)`}
               onChange={this._setTravelTimePercentile}
             />
+
+            <AdvancedSettings
+              addEditRegionalAnalysisBoundsLayerToMap={
+                addEditRegionalAnalysisBoundsLayerToMap
+              }
+              disabled={disableInputs}
+              profileRequest={profileRequest}
+              regionalAnalyses={regionalAnalyses}
+              regionBounds={regionBounds}
+              removeEditRegionalAnalysisBoundsLayerFromMap={
+                removeEditRegionalAnalysisBoundsLayerFromMap
+              }
+              setProfileRequest={setProfileRequest}
+            />
           </div>
         </InnerDock>
       </div>
-    )
-  }
-
-  /** Render the dropdown box that allows selecting regional analysis bounds */
-  renderRegionalAnalysisBoundsSelector (disableInputs: boolean) {
-    const {regionalAnalysisBounds, regionBounds, regionalAnalyses} = this.props
-
-    // figure out which is selected
-    let selected
-    if (
-      !regionalAnalysisBounds ||
-      boundsEqual(regionalAnalysisBounds, regionBounds)
-    ) {
-      selected = '__REGION' // special value
-    } else {
-      const selectedAnalysis = regionalAnalyses.find(r =>
-        boundsEqual(webMercatorBoundsToGeographic(r), regionalAnalysisBounds)
-      )
-
-      if (selectedAnalysis != null) selected = selectedAnalysis._id
-      else selected = '__CUSTOM'
-    }
-
-    const options = [
-      {value: '__REGION', label: messages.analysis.regionalBoundsRegion},
-      ...regionalAnalyses.map(({name, _id}) => ({
-        value: _id,
-        label: sprintf(messages.analysis.regionalBoundsSame, name)
-      }))
-    ]
-
-    if (selected === '__CUSTOM') {
-      // Don't allow the user to select 'Custom' - to make custom bounds, just drag the map markers
-      options.push({
-        value: '__CUSTOM',
-        label: messages.analysis.regionalBoundsCustom,
-        disabled: true
-      })
-    }
-
-    const {editingBounds} = this.state
-
-    return (
-      <Group label={messages.analysis.regionalBounds}>
-        <div className='row'>
-          <div className='col-xs-6'>
-            <Select
-              clearable={false}
-              disabled={disableInputs}
-              value={selected}
-              options={options}
-              onChange={this._setRegionalAnalysisBounds}
-            />
-          </div>
-          <div className='col-xs-6'>
-            {editingBounds
-              ? <Button
-                block
-                onClick={this._hideBoundsSelector}
-                style='warning'
-                >
-                <Icon type='stop' /> Stop editing bounds
-                </Button>
-              : <Button
-                disabled={disableInputs}
-                block
-                onClick={this._showBoundsSelector}
-                style='warning'
-                >
-                <Icon type='pencil' /> Set custom geographic bounds
-                </Button>}
-          </div>
-        </div>
-      </Group>
     )
   }
 
@@ -526,41 +439,5 @@ function ProjectApplicationErrors ({errors}) {
         <ProjectApplicationError error={err} key={`err-${idx}`} />
       ))}
     </div>
-  )
-}
-
-/** Convert web mercator bounds from a regional analysis to geographic bounds */
-function webMercatorBoundsToGeographic ({
-  north,
-  west,
-  width,
-  height,
-  zoom
-}: {
-  north: number,
-  west: number,
-  height: number,
-  width: number,
-  zoom: number
-}) {
-  const nw = LeafletMap.prototype.unproject([west + 1, north], zoom)
-  const se = LeafletMap.prototype.unproject(
-    [west + width + 1, north + height],
-    zoom
-  )
-  return {
-    north: nw.lat,
-    south: se.lat,
-    east: se.lng,
-    west: nw.lng
-  }
-}
-
-function boundsEqual (b0, b1, epsilon = 1e-6) {
-  return (
-    Math.abs(b0.north - b1.north) < epsilon &&
-    Math.abs(b0.west - b1.west) < epsilon &&
-    Math.abs(b0.east - b1.east) < epsilon &&
-    Math.abs(b0.south - b1.south) < epsilon
   )
 }
