+++ lib/styles.css
@@ -404,6 +404,10 @@ label {
   padding-right: 12px;
 }
 
+.advanced-panel {
+  border: 1px solid #ddd !important;
+}
+
 .modification-select.panel-heading {
   padding: 12px 12px 12px 12px;
   cursor: pointer;
@@ -747,4 +751,3 @@ label {
 .JobProgress-delivered {
   background-color: #2389c9;
 }
-
