+++ lib/types.js
@@ -206,15 +206,22 @@ export type Modification = Model & {
 
 export type ProfileRequest = {
   accessModes: string,
-  egressModes: string,
+  bikeSpeed: number,
+  date: string,
   directModes: string,
-  fromTime: number,
-  toTime: number,
+  egressModes: string,
   fromLat: number,
   fromLon: number,
+  fromTime: number,
+  maxBikeTime: number,
+  maxRides: number,
+  maxWalkTime: number,
+  monteCarloDraws: number,
+  toTime: number,
   transitModes: string,
   travelTimePercentile: number,
-  variantIndex: number
+  variantIndex: number,
+  walkSpeed: number
 }
 
 export type Bookmark = Model & {
