+++ lib/utils/mock-data.js
@@ -223,23 +223,28 @@ export const mockModification = {
 }
 
 export const mockProfileRequest = {
-  date: '2016-01-16',
-  fromTime: 25200,
-  toTime: 32400,
   accessModes: 'WALK',
-  directModes: 'WALK',
-  egressModes: 'WALK',
-  transitModes: 'BUS,TRAM,RAIL,SUBWAY',
-  walkSpeed: 1.3888888888888888,
-  bikeSpeed: 4.166666666666667,
-  reachabilityThreshold: 0,
   bikeSafe: 1,
   bikeSlope: 1,
+  bikeSpeed: 4.166666666666667,
   bikeTime: 1,
-  monteCarloDraws: 200,
+  date: '2016-01-16',
+  directModes: 'WALK',
+  egressModes: 'WALK',
+  fromLat: 12,
+  fromLon: 34,
+  fromTime: 25200,
+  maxBikeTime: 20,
   maxRides: 4,
   maxTripDurationMinutes: 60,
-  travelTimePercentile: 50
+  maxWalkTime: 20,
+  monteCarloDraws: 200,
+  reachabilityThreshold: 0,
+  toTime: 32400,
+  transitModes: 'BUS,TRAM,RAIL,SUBWAY',
+  travelTimePercentile: 50,
+  variantIndex: 0,
+  walkSpeed: 1.3888888888888888
 }
 
 export const mockStores = {
